fix: integração das operacoes financeiras ao service central de contas
