+++ front-end/src/app/features/client/services/saque.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, inject } from '@angular/core';
-import { Observable, of, throwError } from 'rxjs';
-import { API_URL } from '../../../core/configs/api.token';
-import { HttpClient } from '@angular/common/http';
+import { Observable, map } from 'rxjs';
+
+import { ContaService } from '../../../core/services/conta.service';
 
 export interface SaqueRequest {
   contaOrigem: string;
@@ -17,11 +17,14 @@ export interface SaqueResponse {
   providedIn: 'root',
 })
 export class SaqueService {
-
-  private readonly http = inject(HttpClient);
-  private readonly apiBaseUrl = inject(API_URL);
+  private readonly contaService = inject(ContaService);
 
   realizarSaque(request: SaqueRequest): Observable<SaqueResponse> {
-    return this.http.post<SaqueResponse>(`${this.apiBaseUrl}/transacoes/saque`, request);
+    return this.contaService.sacar(request.contaOrigem, request.valor).pipe(
+      map((resposta) => ({
+        message: 'Saque realizado com sucesso!',
+        novoSaldoOrigem: resposta.saldo,
+      })),
+    );
   }
 }
